SimicsOpenBoardPkg： Initialize temporary memory with PcdInitValueInTempStack
PeiCore dumps how many bytes of stack is used by checking stack
contents against PcdInitValueInTempStack.
The assumption is when temporary memory is setup, its initial content
is PcdInitValueInTempStack.

The patch changes X64 version SecCore of QSP to fill the temporary
memory as what PeiCore expects.

This helps to detect if stack in temporary ram is enough.

Signed-off-by: Ray Ni <ray.ni@intel.com>